Feature: specify pubsub topics in the config file
Problem: the "alive" and "queue" topics were specified in different
sections of the config file, making it impossible to write
the subscription code generically.

Solution: provide both topics as a list in the config under
the `p2p` section.

Note that the previous values were not removed (yet) to keep
backward compatibility with the CCN config files.