Feature: return 403 when a user dials the wrong peer (#14)
We now make a difference between a dial simply failing and
an error where the user specified the wrong peer ID for
a known multiaddress. When using the /api/p2p/dial endpoint,
the service will now return a 403 response when a conflicting
peer ID is specified, i.e. the peer ID specified by the user
does not match the peer ID of the multiaddress he tries to
dial.