Fix: use correct sender peer ID in routing key
Problem: the "source" part of the routing key for the subscription
exchange was set to the propagation source (i.e. the peer who
forwarded the message) instead of the original sender.

Solution: use the peer ID of the original sender.

Note that theoretically the sender peer ID can be unspecified.
On such a message, the service will now log a warning and drop
the message without propagating it on the message queue.