Feature: list public multiaddresses in /p2p/identify
Problem: users of the service would like to know through which address
the node is accessible.
Solution: expose the external addresses of the P2P swarm in
the /p2p/identify endpoint.

Added a new command, `Identify`, that returns the local peer ID
and the list of external addresses of the swarm.