Fix: avoid reading RabbitMQ messages without ack-ing them (#29)
Problem: the P2P service hangs randomly because a message was not
acknowledged within 30 minutes of being fetched for the pub queue.

Solution: run the P2P <-> MQ tasks in two separate tasks instead
of using `tokio::select!`.

`tokio::select!` cancels the future(s) that did not complete.
This is the probable cause of the issue. Introduced two new tasks
to replace the `p2p_loop` task so that they run uninterrupted.