Fix: support concurrent P2P API calls (#55)
Problem: API calls are serialized because we use the mutex around the
P2P client object a bit too generously in the API.

Solution: lock the mutex for a shorter amount of time. We now lock the
mutex for the time needed to send the command to the P2P background task
and then wait for the result after unlocking the mutex.